Merged manually the Total Angular Momentum method which was implemented in a separate branch in 2020. The test_class.py has been modified accordingly.
It passes all the nosetests

Results are slightly different from the devel branch but I have checked that this is due to the previously incorrect TCA for tensor modes.
I have checked on plots that the new TCA is correct since it leads to a smooth transition when TCA is turned off, whereas it was
not the case before.