Vector modes have been implemented consistently in both the optimal and the TAM hierarchy (except for ncdm where only optimal is used). This required to modify input.c perturbations.c transfer.c harmonic.c primordial.c and output.c (and their .h files) along with precision.h
The structure is mostly inherited from tensor modes, and the approximations schemes are TCA, full hierarchies, RSA.

The neutrinos and ncdm are integrated even in RSA, as is the case currently for tensor modes. This could be removed to improve greatly the speed.

The transfer_radial_function() had a typo for vector modes which I have corrected.

The vector metric perturbation is the same in synchronous and newtonian gauge. The gauge is rather a change of frame for velocities and dipole with
v^synchronous = v^newtonian - V
where V is the metric vector perturbation

The initial conditions implemented are the standard isocurvature modes (where the velocity of neutrinos and photons is somehow opposite initially) as in 0403583.

I have also added a fancy neutrino octupole condition which is compatible with a regular metric perturbation.

The test_class.py has been extended to allow testing of vectors (but only with the isocurvature initial condition).
I am not sure what parameter must go in normal and in power given the peculiarity of the vector modes.

The consistency between the multipoles as integrated by the hierarchy, and those from the line-of-sight has been checked (which is how the typo in radial_function was spotted).

The tight-coupling approximations at early time have been checked by plotting and checking the smooth transition.