Since vector modes have been implemented, it must also be possible for the user to choose to output the transfer functions of vector modes (the velocities and the vector metric potential) I have added an option 'wTk' for outputs, (w stands here for a common notation for vorticity, since vector modes of velocities are essentially vorticity).
It required to modify as well the output functions, so that they depend on index_md, and the scalar and vector cases are split inside them with
if (_scalars_)
if (_vectors_)

Also since the neutrinos were slowing the code a lot at late time, they are now removed in the RSA approximation.
This does not affect the results.

Maybe one should think of doing it for tensor perturbations since neutrinos affect the GW subdominantly and could be removed in RSA instead of slowing the code.